refactor(log): move context to pkg
