+++ internal/api/api.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"nimona.io/internal/http/router"
-	"nimona.io/internal/log"
+	"nimona.io/pkg/log"
 	"nimona.io/internal/store/graph"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/discovery"
