+++ pkg/exchange/exchange.go
@@ -12,7 +12,7 @@ import (
 
 	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
-	"nimona.io/internal/log"
+	"nimona.io/pkg/log"
 	"nimona.io/internal/store/graph"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/discovery"
