+++ pkg/log/writer_string.go
