+++ pkg/middleware/handshake/handshake.go
@@ -2,7 +2,7 @@ package handshake
 
 import (
 	"nimona.io/pkg/context"
-	"nimona.io/internal/log"
+	"nimona.io/pkg/log"
 	"nimona.io/internal/rand"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/discovery"
