+++ pkg/net/network.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 
 	"nimona.io/internal/errors"
-	"nimona.io/internal/log"
+	"nimona.io/pkg/log"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/discovery"
 	"nimona.io/pkg/peer"
