+++ pkg/orchestrator/orchestrator_sync.go
@@ -3,7 +3,7 @@ package orchestrator
 import (
 	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
-	"nimona.io/internal/log"
+	"nimona.io/pkg/log"
 	"nimona.io/pkg/exchange"
 )
